// src/model/interfaces/estudiante/infoEstudiante.ts
export interface InfoEstudiante {
    est_Id: string
    estudiante: string
    email1: string
    email2: string
    celular: string
    programa: string
    sede: string
    facultad: string
    carrera: string
    planEstudio: string
    asi_Id: number
    asignatura: string
    anioActual: string
    periodoActual: string
    periodoId: number
}